Restrict options within subcommands, begin building API
It is obvious that the following are equivalent:

    example -xyz  subcommand <files>...
    example -xyz (subcommand <files>...)

However, some arrangements should not be legal:

    example -xy subcommand -z file1 file2

because it would appear that the -z option is associated with the subcommand,
though the usage patterns would imply otherwise. This is remedied by tracking
the options that are used while creating the tree; a subcommand will push and
empty tracking onto the stack of trackers and a symbolic node is added to pop
the tracker off the stack.

Currently, brackets, parens, and pipes are not supported (though parens should
be added soon).